% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isothermal_models.R
\name{Arrhenius_iso}
\alias{Arrhenius_iso}
\title{Isothermal Arrhenius model}
\usage{
Arrhenius_iso(time, temp, k_ref, temp_ref, Ea)
}
\arguments{
\item{time}{numeric indicating the time at which the prediction is taken.}

\item{temp}{numeric indicating the temperature of the treatment.}

\item{k_ref}{numeric indicating the inactivation rate at the reference
temperature.}

\item{temp_ref}{numeric indicating the reference temperature.}

\item{Ea}{numeric indicating the activation energy.}
}
\value{
A numeric with the predicted logarithmic reduction
(\eqn{log10(N/N0)}).
}
\description{
Returns the predicted logarithmic reduction in microbial count according
to Arrhenius model for the time, temperature and model parameters
given.
}
