% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isothermal_models.R
\name{Bigelow_iso}
\alias{Bigelow_iso}
\title{Isothermal Bigelow's Model}
\usage{
Bigelow_iso(time, temp, z, D_R, temp_ref)
}
\arguments{
\item{time}{numeric indicating the time at which the prediction is taken.}

\item{temp}{numeric indicating the temperature of the treatment.}

\item{z}{numeric defining the z-value.}

\item{D_R}{numeric defining the D-value at the reference temperature.}

\item{temp_ref}{numeric defining the reference temperature.}
}
\value{
A numeric with the predicted logarithmic reduction
(\eqn{log10(N/N0)}).
}
\description{
Returns the predicted logarithmic reduction in microbial count according
to Bigelow's model for the time, temperature and model parameters given.
}
