% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_examples.R
\docType{data}
\name{dynamic_inactivation}
\alias{dynamic_inactivation}
\title{Example Dynamic Inactivation of a Microorganis}
\format{
A data frame with 19 rows and 2 variables.
}
\usage{
data(dynamic_inactivation)
}
\description{
Example of experimental data of the dynamic inactivation process of a
microorganism.
}
\details{
\itemize{
\item time: Time in minutes of the measurement.
\item N: Number of microorganism.
\item temperature: Observed temperature.
}
}
\keyword{datasets}
