% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isothermal_models.R
\name{Geeraerd_iso}
\alias{Geeraerd_iso}
\title{Isothermal Geeraerd Model}
\usage{
Geeraerd_iso(time, temp, logC0, a, z)
}
\arguments{
\item{time}{numeric vector with the treatment time}

\item{temp}{numeric vector with the treatment temperature}

\item{logC0}{model parameter describing the shoulder length}

\item{a}{model parameter describing the intercept of the relation}

\item{z}{z-value}
}
\description{
Returns the predicted logarithmic reduction in microbial cont according
to Geeraerd's model.
The isothermal prediction is calculated by analytical integration of the
ode for constant temperature
}
