% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isothermal_fit.R
\name{get_isothermal_model_data}
\alias{get_isothermal_model_data}
\title{Isothermal Model Data}
\usage{
get_isothermal_model_data(model_name = "valids")
}
\arguments{
\item{model_name}{Optional string with the key of the model to use.}
}
\value{
If \code{model_name} is missing, a list of the valid model keys.
If \code{model_name} is not a valid key, NULL is returned.
Otherwise, a list with the parameters of the model selected and its
\code{formula} for the nonlinear adjustment.
}
\description{
Provides information of the models implemented for fitting of isothermal
data.
This models are valid only for isothermal adjustment with the function
\code{\link{fit_isothermal_inactivation}}. To make predictions with the
function \code{\link{predict_inactivation}} or adjust dynamic experiments
with \code{\link{fit_dynamic_inactivation}}, use
\code{\link{get_model_data}}.
}
