% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_functions.R
\name{goodness_dyna}
\alias{goodness_dyna}
\title{Goodness of fit for Dynamic fits}
\usage{
goodness_dyna(object)
}
\arguments{
\item{object}{An instance of FitInactivation}
}
\description{
Goodness of fit for Dynamic fits
}
