% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_functions.R
\name{goodness_iso}
\alias{goodness_iso}
\title{Goodness of fit for Isothermal fits}
\usage{
goodness_iso(object)
}
\arguments{
\item{object}{An object of class IsoFitInactivation.}
}
\description{
Goodness of fit for Isothermal fits
}
