% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_functions.R
\name{goodness_MCMC}
\alias{goodness_MCMC}
\title{Goodness of fit for MCMC fits}
\usage{
goodness_MCMC(object)
}
\arguments{
\item{object}{An instance of FitInactivationMCMC}
}
\description{
Goodness of fit for MCMC fits
}
