% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC_functions.R
\name{goodness_of_fit}
\alias{goodness_of_fit}
\title{Goodness of fit for microbial inactivation models}
\usage{
goodness_of_fit(object)
}
\arguments{
\item{object}{A model fit generated by bioinactivation}
}
\description{
Generates a table with several statistical indexes describing the
quality of the model fit:
\itemize{
\item ME: Mean Error.
\item RMSE: Root Mean Squared Error.
\item loglik: log-likelihood.
\item AIC: Akaike Information Criterion.
\item AICc: Akaike Information Criterion with correction for
finite sample size.
\item BIC: Bayesian Infromation Criterion.
\item Af: Accuracy factor.
\item Bf: Bias factor.
}
}
