% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3_others.R
\name{is.FitInactivation}
\alias{is.FitInactivation}
\title{Test of FitInactivation object}
\usage{
is.FitInactivation(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A logic specifying whether \code{x} is of class
\code{FitInactivation}
}
\description{
Tests if an object is of class \code{FitInactivation}.
}
