% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3_others.R
\name{is.FitInactivationMCMC}
\alias{is.FitInactivationMCMC}
\title{Test of FitInactivationMCMC object}
\usage{
is.FitInactivationMCMC(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A logic specifying whether \code{x} is of class
\code{FitInactivationMCMC}
}
\description{
Tests if an object is of class \code{FitInactivationMCMC}.
}
