% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3_others.R
\name{is.IsoFitInactivation}
\alias{is.IsoFitInactivation}
\title{Test of IsoFitInactivation object}
\usage{
is.IsoFitInactivation(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A logic specifying whether \code{x} is of class
\code{IsoFitInactivation}
}
\description{
Tests if an object is of class \code{IsoFitInactivation}.
}
