% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3_others.R
\name{is.PredInactivationMCMC}
\alias{is.PredInactivationMCMC}
\title{Test of PredInactivationMCMC object}
\usage{
is.PredInactivationMCMC(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A logic specifying whether \code{x} is of class
\code{PredInactivationMCMC}
}
\description{
Tests if an object is of class \code{PredInactivationMCMC}.
}
