% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3_others.R
\name{is.SimulInactivation}
\alias{is.SimulInactivation}
\title{Test of SimulInactivation object}
\usage{
is.SimulInactivation(x)
}
\arguments{
\item{x}{object to be checked.}
}
\value{
A logic specifying whether \code{x} is of class
\code{SimulInactivation}
}
\description{
Tests if an object is of class \code{SimulInactivation}.
}
