% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_examples.R
\docType{data}
\name{isothermal_inactivation}
\alias{isothermal_inactivation}
\title{Example Isothermal Inactivation of a Microorganis}
\format{
A data frame with 36 rows and 3 variables.
}
\usage{
data(isothermal_inactivation)
}
\description{
Example of experimental data for an isothermal process of a
microorganism.
}
\details{
\itemize{
\item time: Time in minutes of the measurement.
\item temp: Temperature at which the experiment was made.
\item log_diff: Logarithmic difference.
}
}
\keyword{datasets}
