% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_examples.R
\docType{data}
\name{laterosporus_dyna}
\alias{laterosporus_dyna}
\title{Example Dynamic Inactivation of a Laterosporus}
\format{
A data frame with 20 rows and 3 variables.
}
\usage{
data(laterosporus_dyna)
}
\description{
Example of experimental data of the dynamic inactivation process of
Laterosporus
}
\details{
\itemize{
\item time: Time in minutes of the measurement.
\item temp: observed temperature.
\item logN: recorded number of microorganism.
}
}
\keyword{datasets}
