% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_examples.R
\docType{data}
\name{laterosporus_iso}
\alias{laterosporus_iso}
\title{Example Isothermal Inactivation of a Laterosporus}
\format{
A data frame with 52 rows and 3 variables.
}
\usage{
data(laterosporus_iso)
}
\description{
Example of experimental data for an isothermal process of
Laterosporus.
}
\details{
\itemize{
\item time: Time in minutes of the measurement.
\item temp: Temperature at which the experiment was made.
\item log_diff: Logarithmic difference.
}
}
\keyword{datasets}
