% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3_plot.R
\name{plot.FitInactivationMCMC}
\alias{plot.FitInactivationMCMC}
\title{Plot of FitInactivationMCMC Object}
\usage{
\method{plot}{FitInactivationMCMC}(x, y = NULL, ..., make_gg = TRUE)
}
\arguments{
\item{x}{the object of class \code{FitInactivation} to plot.}

\item{y}{ignored}

\item{...}{additional arguments passed to \code{plot}.}

\item{make_gg}{logical. If \code{TRUE}, the plot is created using
\code{ggplot2}. Otherwise, the plot is crated with base \code{R}.
\code{TRUE} by default.}
}
\value{
If \code{make_gg = FALSE}, the plot is created. Otherwise, an
        an instance of \code{ggplot} is generated, printed and returned.
}
\description{
Plots a comparison between the experimental data provided and the prediction
produced by the model parameters adjusted for an instance of
\code{FitInactivationMCMC}.
}

