% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3_plot.R
\name{plot.PredInactivationMCMC}
\alias{plot.PredInactivationMCMC}
\title{Plot of PredInactivationMCMC Object}
\usage{
\method{plot}{PredInactivationMCMC}(x, y = NULL, ..., make_gg = TRUE)
}
\arguments{
\item{x}{the object of class \code{PredInactivationMCMC} to plot.}

\item{y}{ignored}

\item{...}{additional arguments passed to \code{plot}.}

\item{make_gg}{logical. If \code{TRUE}, the plot is created using
\code{ggplot2}. Otherwise, the plot is crated with base \code{R}.
\code{TRUE} by default.}
}
\value{
If \code{make_gg = FALSE}, the plot is created. Otherwise, an
an instance of \code{ggplot} is generated, printed and returned.
}
\description{
Plots the prediction interval generated by
\code{\link{predict_inactivation_MCMC}}.
}
\details{
The plot generated in ggplot (default) generates a dashed line with the mean
of the MC
simulations. Moreover, a ribbon with the 2 first quantiles (i.e. columns 3
and 4) is generated.

The plot generated with base R (make_gg = \code{FALSE}) generates a solid line
wth the mean of the MC simulations. Each one of the other quantiles included
in the data frame are added with different
}
