% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3_plot.R
\name{plot.SimulInactivation}
\alias{plot.SimulInactivation}
\title{Plot of SimulInactivation Object}
\usage{
\method{plot}{SimulInactivation}(x, y = NULL, ..., make_gg = TRUE)
}
\arguments{
\item{x}{The object of class \code{SimulInactivation} to plot.}

\item{y}{ignored}

\item{...}{additional arguments passed to \code{plot}.}

\item{make_gg}{logical. If \code{TRUE}, the plot is created using
\code{ggplot2}. Otherwise, the plot is crated with base \code{R}.
\code{TRUE} by default.}
}
\value{
If \code{make_gg = FALSE}, the plot is created. Otherwise, an
        an instance of \code{ggplot} is generated, printed and returned.
}
\description{
Plots the predicted evolution of the logarithmic count with time for an
instance of \code{SimulInactivation}.
}

