% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3_others.R
\name{summary.FitInactivation}
\alias{summary.FitInactivation}
\title{Summary of a FitInactivation object}
\usage{
\method{summary}{FitInactivation}(object, ...)
}
\arguments{
\item{object}{Instance of Fit Inactivation}

\item{...}{ignored}
}
\description{
Summary of a FitInactivation object
}
