% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3_others.R
\name{summary.FitInactivationMCMC}
\alias{summary.FitInactivationMCMC}
\title{Summary of a FitInactivationMCMC object}
\usage{
\method{summary}{FitInactivationMCMC}(object, ...)
}
\arguments{
\item{object}{Instance of FitInactivationMCMC}

\item{...}{ignored}
}
\description{
Summary of a FitInactivationMCMC object
}
