% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3_others.R
\name{summary.IsoFitInactivation}
\alias{summary.IsoFitInactivation}
\title{Summary of a IsoFitInactivation object}
\usage{
\method{summary}{IsoFitInactivation}(object, ...)
}
\arguments{
\item{object}{Instance of IsoFitInactivation}

\item{...}{ignored}
}
\description{
Summary of a IsoFitInactivation object
}
