% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods_s3.R
\name{plot.FitInactivationMCMC}
\alias{plot.FitInactivationMCMC}
\title{Plot of FitInactivationMCMC Object}
\usage{
\method{plot}{FitInactivationMCMC}(x, y = NULL, ...)
}
\arguments{
\item{x}{the object of class \code{FitInactivation} to plot.}

\item{y,...}{ignored}
}
\description{
Plots a comparison between the experimental data provided and the prediction
produced by the model parameters adjusted for an instance of
\code{FitInactivationMCMC}.
}

