% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods_s3.R
\name{plot.SimulInactivation}
\alias{plot.SimulInactivation}
\title{Plot of SimulInactivation Object}
\usage{
\method{plot}{SimulInactivation}(x, y = NULL, ...)
}
\arguments{
\item{x}{The object of class \code{SimulInactivation} to plot.}

\item{y,...}{ignored}
}
\description{
Plots the predicted evolution of the logarithmic count with time for an
instance of \code{SimulInactivation}.
}

