% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_s3_plot.R
\name{plot.SimulInactivation}
\alias{plot.SimulInactivation}
\title{Plot of SimulInactivation Object}
\usage{
\method{plot}{SimulInactivation}(
  x,
  y = NULL,
  ...,
  make_gg = TRUE,
  plot_temp = FALSE,
  label_y1 = "logN",
  label_y2 = "Temperature",
  ylims = NULL
)
}
\arguments{
\item{x}{The object of class \code{SimulInactivation} to plot.}

\item{y}{ignored}

\item{...}{additional arguments passed to \code{plot}.}

\item{make_gg}{logical. If \code{TRUE}, the plot is created using
\code{ggplot2}. Otherwise, the plot is crated with base \code{R}.
\code{TRUE} by default.}

\item{plot_temp}{logical. Whether the temperature profile will
be added to the plot. \code{FALSE} by default.}

\item{label_y1}{Label of the principal y-axis.}

\item{label_y2}{Label of the secondary y-axis.}

\item{ylims}{Numeric vector of length 2 with the Limits of the
y-axis. \code{NULL} by default (0, max_count).}
}
\value{
If \code{make_gg = FALSE}, the plot is created. Otherwise, an
an instance of \code{ggplot} is generated, printed and returned.
}
\description{
Plots the predicted evolution of the logarithmic count with time for an
instance of \code{SimulInactivation}.
}
